Create README.md
HapticCompass app is an Android application created for my Masters IoT project at the University of Glasgow. The aim of project was to explore possibilities of creating a prototype, which could use Android powered device build-in sensors to drive haptic sensations. The app provides basic compass functionality that in real time sends sensors information to The Particle Cloud.